+++ packages/server-asset/tsup.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/index.ts", "src/hooks/data.ts", "src/hooks/wasm.ts"],
+  format: ["esm"],
+  dts: true,
+});